backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
pyodbc==4.0.39
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
google-auth==2.23.3
pandas==2.1.2
python-dotenv==1.0.0
python-magic==0.4.27
numpy==1.26.1
openpyxl==3.1.2
aiofiles==23.2.1